Adminimal sub-theme with media archiving on old node revisions
If a media entity is attached to a published version of a node entity, the file is moved into the public file system. Anonymous users can then access the file.

If a media entity not attached to a published version of a node entity, the file is moved into the private file system and the media entity is unpublished. This prevents access from anonymous users.

The trigger is a submit handler on the node edit form.

Only applies to 2 new media types, Managed files and Managed images.

To do: 

 1. Support paragraphs
 1. Reverting revisions
 1. Bulk updating (may not be possible)
 1. Support the moderation block form
 1. Update test to include